+++ index.html
@@ -6,45 +6,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Keyboard</title>
     <link rel="stylesheet" href="oldstyle.css">
-    <link rel="icon" href="./assets/favicon.ico">
+    <link rel="icon" href="favicon.ico">
 </head>
 <body>
-    <div class="container">
-        <h1 class="title">Virtual keyboard RS</h1>
-        <textarea class="field" id="field"></textarea>
-        <div class="keyboard">
-            <div class="keyboardRow">
-                <div class="graykey key" id="Digit1">
-                    <span class="eng">
-                        <span class="caseDown">1</span>
-                        <span class="caseUp">!</span>
-                    </span>
-                </div>
-                <div class="graykey key" id="Digit2">
-                    <span class="eng">
-                        <span class="caseDown">2</span>
-                        <span class="caseUp">@</span>
-                    </span>
-                </div>
-                <div class="graykey key" id="Digit3">
-                    <span class="eng">
-                        <span class="caseDown">3</span>
-                        <span class="caseUp">#</span>
-                    </span>
-                </div>
-            </div>
-            <div class="keyboardRow">
-                <div class="darkkey key">
-                    <span class="tab">Tab</span>
-                </div>
-            </div>
-            <div class="keyboardRow"></div>
-            <div class="keyboardRow"></div>
-            <div class="keyboardRow"></div>
-        </div>
-        <p class="description">Клавиатура создана в операционной системе Windows</p>
-        <p class="language">Для переключения языка комбинация: левыe ctrl + alt</p>
-    </div>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 </html>
