--- indexGen.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Keyboard</title>
-    <link rel="stylesheet" href="oldstyle.css">
-    <link rel="icon" href="favicon.ico">
-</head>
-<body>
-    <script type="module" src="indexGen.js"></script>
-</body>
-</html>
