+++ package.json
@@ -1,0 +1,23 @@
+{
+  "name": "virtual_keyboard",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "build": "webpack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Jeromkeee/Virtual_keyboard.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Jeromkeee/Virtual_keyboard/issues"
+  },
+  "homepage": "https://github.com/Jeromkeee/Virtual_keyboard#readme",
+  "devDependencies": {
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2"
+  }
+}
